backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
openai==1.10.0
anthropic==0.18.0
google-generativeai==0.3.2
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
requests==2.31.0
chart.js==4.4.1
reportlab==4.0.9
pandas==2.2.0
numpy==1.26.3